// frontend/src/apps/pos/api/schemas.ts
import { z } from "zod";
import { DecimalSchema } from "@core/schemas/zod-schemas";

export const UserSchema = z.object({
  id: z.number(),
  username: z.string(),
  email: z.string(),
  admin: z.boolean(),
});

export type User = z.infer<typeof UserSchema>;

export const ProductSchema = z.object({
  id: z.number(),
  name: z.string(),
  netPrice: DecimalSchema,
  grossPrice: DecimalSchema,
  vatRate: DecimalSchema,
  vatAmount: DecimalSchema,
  wrapAfter: z.boolean(),
  hidden: z.boolean(),
  soldOut: z.boolean(),
  apiExport: z.boolean(),
  pos: z.number(),
  totalStock: z.number(),
  guestlists: z
    .array(
      z.object({
        id: z.number(),
        name: z.string(),
        typeCode: z.boolean(),
        productId: z.number(),
      }),
    )
    .nullable(),
  unitsSold: z.number(),
  soldOutRequestCount: z.number(),
});

export type Product = z.infer<typeof ProductSchema>;

export const PurchaseItemSchema = z.object({
  id: z.number(),
  purchaseID: z.uuid(),
  productID: z.number(),
  product: ProductSchema,
  quantity: z.number(),
  netPrice: DecimalSchema,
  grossPrice: DecimalSchema,
  vatRate: DecimalSchema,
  vatAmount: DecimalSchema,
  totalNetPrice: DecimalSchema,
  totalGrossPrice: DecimalSchema,
  totalVatAmount: DecimalSchema,
});

export type PurchaseItem = z.infer<typeof PurchaseItemSchema>;

const PurchaseStatusSchema = z.enum([
  "pending",
  "confirmed",
  "refunded",
  "failed",
]);

export const PurchaseSchema = z.object({
  id: z.uuid(),
  createdAt: z.string(),
  createdBy: z.string(),
  paymentMethod: z.string(),
  totalNetPrice: DecimalSchema,
  totalGrossPrice: DecimalSchema,
  totalVatAmount: DecimalSchema,
  sumupTransactionId: z.uuid().nullable(),
  sumupClientTransactionId: z.uuid().nullable(),
  status: PurchaseStatusSchema,
  purchaseItems: z.array(PurchaseItemSchema),
});

export type Purchase = z.infer<typeof PurchaseSchema>;

export const GuestSchema = z.object({
  id: z.number(),
  name: z.string(),
  code: z.string().nullable(),
  listName: z.string(),
  additionalGuests: z.number().min(0).max(10),
  arrivalNote: z.string().nullable(),
  attendedGuests: z.number().default(0),
});

export type Guest = z.infer<typeof GuestSchema>;

export const ProductInterestSchema = z.object({
  id: z.number(),
});

export type ProductInterest = z.infer<typeof ProductInterestSchema>;